refactor(apm-builder): extract matchesGlob into lib/glob.ts
