+++ apm-builder/cli.ts
@@ -5,7 +5,8 @@ import pc from 'picocolors';
 import chokidar from 'chokidar';
 import { discoverComponents } from './lib/discover.ts';
 import { validateComponents } from './lib/validate.ts';
-import { runBuild, matchesGlob } from './lib/build.ts';
+import { runBuild } from './lib/build.ts';
+import { matchesGlob } from './lib/glob.ts';
 import { renderReadme } from './lib/docs.ts';
 import { TARGETS, type Target } from './lib/types.ts';
 import { listImplementedTargets } from './adapters/index.ts';
