+++ apm-builder/lib/glob.ts
@@ -1,0 +1,4 @@
+export function matchesGlob(name: string, pattern: string): boolean {
+  const re = new RegExp('^' + pattern.replace(/[.+?^${}()|[\]\\]/g, '\\$&').replace(/\*/g, '.*') + '$');
+  return re.test(name);
+}
